cscvs to tla changeset 4
Author: srivasta
Date: 1997/06/03 03:47:53
Added RCS keywords. Made sure that the default image is zImage for ppc
and powerpc architectures (the former is incorrect, but was reported
by some versions of dpkg), and is vmlinux for sparc and m68k
architectures. Also made sure that the rules do not attempt to install
a compressed kernel image in the latter cases.

git-archimport-id: srivasta@debian.org--2003-primary/kernel-package--devel--0.1--patch-3