cscvs to tla changeset 319
Author: srivasta
Date: 2000/03/03 16:03:55
   * Path up modules_install in the kernel_image target to include some new
module categories in 2.3.XX kernels. In order that this does not, in
the futire, inconvenience people, use the kernel Makefile to install
modules when we can. The cases we can't use the kernel makefile is for
2.0.XX and older kernels (since the prefix to modlib was not present
then). closes: Bug#59516
* Fixed some typos. Since the bug report did ot indicate what the typos
were, and I am  not inclined to play guessing games, I am assuming the
typos fixed were what the reporter intended. closes: Bug#59528

git-archimport-id: srivasta@debian.org--2003-primary/kernel-package--devel--0.1--patch-318