cscvs to tla changeset 349
Author: srivasta
Date: 2000/07/11 04:25:33
   * Added a new configuration directive for the target machine, namely,
realtive_links. Normally, the kernel image postinst shall do a
relative link if possible, or else it shall make an absolute
link. With this directive, the postinst shall go to extraordinary
lengths to make sure that the link is relative. The previous release
also closed a but, namely: closes: Bug#65116

git-archimport-id: srivasta@debian.org--2003-primary/kernel-package--devel--0.1--patch-348