cscvs to tla changeset 363
Author: srivasta
Date: 2001/01/19 19:35:09
   * For folks using grub, there are now postinst_hook and postrm_hook
variables that can be pointed to scripts that add or remove a line
from the grub menu list at kerel image install and remove times. A
sample script to add lines to a grub menu file is included in the dir
/usr/share/doc/kernel-pacakage/. The script, called
kernel_grub_conf.sh, is coutesy of Junichi Uekawa <dancer@debian.org>
with minor modifications by yours truly.
* Added a postrm_hook variable that can be used to add a script to run
when removing a kernel image package

git-archimport-id: srivasta@debian.org--2003-primary/kernel-package--devel--0.1--patch-362