cscvs to tla changeset 459
Author: srivasta
Date: 2002/03/20 05:21:08
   * Hmm. There was a bug report that said kernel-package does not call
make correctly when concurrency is set; however, this happens to be a
mischaracterization. The top level make call a recursive make: $(MAKE)
-j2; the level 2 make then calls another sub make, and the level 3
make spits out the error that it is running with -j1, as documneted. I
have cleaned up soe of the rules involved to get greater transparency
in the phenomena. Also, remove benign warning about not finding the
MODULES_LOC dir.                                         closes: Bug#139132
* Make the bash completions file go in the right place.

git-archimport-id: srivasta@debian.org--2003-primary/kernel-package--devel--0.1--patch-458