cscvs to tla changeset 488
Author: srivasta
Date: 2002/09/01 01:15:03
   * Hack ake-kpkg so that a clean command by itself always sets debian
revision, optionally reading from the changelog if it exists. This is
a hack meant for people who have set options to make the debian
revisions mandatory, but call clean on a pristince source directory.
This then overrides whatever is in the conf file, but that should not
matter, since we are just going to clean anyway.         closes: Bug#155749
* Provide for means of setting up hook scripts that are run while
packaging the sourc, and can poiunt to executable that massage the
source tree before packaging. Can be used, for example, to get rid of
version control trees, rune away unwanted architectures, etc.
closes: Bug#157128
* Clarified the option inage_on_boot. The new name is going to be
link_in_boot.

git-archimport-id: srivasta@debian.org--2003-primary/kernel-package--devel--0.1--patch-487