cscvs to tla changeset 509
Author: srivasta
Date: 2003/03/22 18:59:09
   * Export MODDIR to modules so that they have an idea where to put the
final debs, rather than just $(KSRC)/... This allows people building
the .debs to be able to specify where the modules .debs go to. Also
mentioned all the variables that are set in README.modules, since that
file was out of date.                                    closes: Bug#185279
* Also add MODULE_LOC to the list of ENVIRONMENT VARIABLES in the
make-kpkg man page (it was already documented in the manual page, just
not in the list)                                         closes: Bug#185859

git-archimport-id: srivasta@debian.org--2003-primary/kernel-package--devel--0.1--patch-508