cscvs to tla changeset 510
Author: srivasta
Date: 2003/03/25 05:03:09
   * Re named MODDIR to KDIST_DEST_DIR to avoid name space clashed.
closes: Bug#185279
* Fix settin up OOT_CMD                                    closes: Bug#186177
* Also pass in $(maintainer) to the modules, since this could be
different from $(pgp).  This is for Bug#186187
* Due to dependencies, 'make-kpkg buildpackage' results in the kernel
being patched, configured, unpatched, cleaned, repatched and
configured again before building starts.  Adding the -nc options to
the commandline for dpkg-buildpackage would prevent it from calling
'debian/rules clean' before it calls 'debian/rules binary'.  To insure
the package is in a sane state if it was previously built, make
buildpackage rely on 'clean' (this appears safe for a fresh kernel
which does not even have the debian directory yet). Most of these
fixes are thanks to chirik@castlefur.com.                closes: Bug#186190

git-archimport-id: srivasta@debian.org--2003-primary/kernel-package--devel--0.1--patch-509