cscvs to tla changeset 533
Author: srivasta
Date: 2003/08/08 19:13:57
   * This patch makes it possible to do things like --subarch=sparc64-smp,
and --subarch=sparc-myhacks, and the rules file will still know that
sparc64 == ARCH==sparc64, etc.                    closes: Bug#203936
* Test the UTS_RELEASE in include/linux/version.h, and touch the
Makefile if it does not match what we think the version is, so that
the version.h file shall be rebuilt.              closes: Bug#204592
* Add parentheses to fix bug cause by ! binding tightly in the check for
the existence of version.h added last upload.     closes: Bug#204173
* Fixed the name of spanish manpages. Reverted all pages to ISO-8859-1,
since otherwise the spanish characters are broken closes: Bug#204365

git-archimport-id: srivasta@debian.org--2003-primary/kernel-package--devel--0.1--patch-532