cscvs to tla changeset 546
Author: srivasta
Date: 2003/08/25 21:00:46
   * Add a postinst hook for source and header images as well. This allows
people to create scripts to do whatever they feel like doing when
headers are instralled, including changing build symlinks if they
wish.                                               closes: Bug#207145
* Remove warnings if not in a kernel directory, but in a kernel-headers
created directory, and if the only target command is a module related
one. This is a massively reworked variation of the patch in the
report, and should have the same functionality.     closes: Bug# 207194

git-archimport-id: srivasta@debian.org--2003-primary/kernel-package--devel--0.1--patch-545