cscvs to tla changeset 554
Author: srivasta
Date: 2003/09/21 19:23:34
   * Added handling of the /lib/modules/`uname -r`/build symbolic link, if
it is dangling. A link to the corresponding kernel-headers location is
added if it exists. One can also force not checking for the existence
of the kernel-headers link.                         closes: Bug#211949
* In some cases, the symlink creation code created a dangling relative
link, instead of an absolute, non dangling link, since the test
relative code only looked at the directory names of the source, rather
than testing for the actual file we needed to link to.
closes: Bug#211140
* We used to silently ignore individual patches and unpatches that were
not found from the list of user requested patches; only complaining if
none of the user requested patches were found. Now we stop early
(before doing any patching at all) if any of the requested patches
can't be found.                                     closes: Bug#211406
* No longer recommend debianutils, since it is an essential package.
closes: Bug#210686

git-archimport-id: srivasta@debian.org--2003-primary/kernel-package--devel--0.1--patch-553