cscvs to tla changeset 557
Author: srivasta
Date: 2003/09/28 01:16:52
   * Install additional useful documentation files in the kernel-doc
package.                                             closes: Bug#212848
* Added more voluble informational messages trying to track down
Bug#212965. Removed Quotes from around the UTS_RELEASE_VERSION
closes: Bug#212965
* Added an option to make-kpkg to allow people to specify the command to
build an initrd image. This is useful if there are many kernel images
in the system with different initrd filesystem. Also added an option
to kernel-img.conf to over ride that variable on a per target basis.
closes: Bug#212602
* Ship man pages and html documentation of the kernel in the kerel-doc
packages. Note that you need to have docbook-utils installed to take
advantage of this.                                   closes: Bug#210630

git-archimport-id: srivasta@debian.org--2003-primary/kernel-package--devel--0.1--patch-556