A bug fix release
There are a whole slew of bugs fixed in this revision. We changed the
recommendation of docbook-utils to a recommends, since formatting the
docbook documentation is a minor part of kernel-package functionality,
and the docbook packages are huge, and may not be relevant to most
users. There were the usual typographical fixes in the manual page. 
We added a new subarch in this revision, namely, PrPMC 750/PrPMC 800. 
Also, we now bug out with a diagnostic if kimagesrc or kimagedest were
undefined; typically the case if we can't determine the arch. Finally,
fixed a bug in the image post installation script so we do create
symbolic links when asked to do so. 

git-archimport-id: srivasta@debian.org--2003-primary/kernel-package--devel--0.2--patch-1