added the ability to package basre vmlinux optionally, and also Kconfig files
Make kernel-package not stop if the revision contains a hypen unless
the said revision fails to meet policy. 

Now kernel-package shall find and package all the configuration
documentation scattered all over the kernel source in 2.6
kernels. This functionality is only invoked for the 2.6 series
kernels, and we delete any empty files (in case there was no
documentation found).

Fixed a couple of typos in error messages, as well as an over zealous
escapes of hyphens in the make-kpkg man page. 

Finally, added a configuration file option that allows one to include
the uncompressed ./vmlinux in the package (in /boot). 



git-archimport-id: srivasta@debian.org--2003-primary/kernel-package--devel--0.2--patch-41