A bug fix release
Went and removed the .TT macro from the man pages, replacing it with
.I. This is suboptimal., itaslics are not tele-type fonts. Fixed a
couple of minor bugs in the wording of the man pages as well. Updated
the spanish translation. 

Also patched the sources to remove XSIisms from the maintainers
scripts and rules files, and make sure that modules clean addresses
all valid modules, not just those in the default location.


git-archimport-id: srivasta@debian.org--2003-primary/kernel-package--devel--0.2--patch-46