Fixed problems with patch application and unapplication
There was a serious problem regarding applying patches where the
patrch list was not provided on the command line, but instead all
installed patches were supposed to be applied (and unapplied). The
basic problem was that the return status of run-parts was being
ignored; the deeper issue was that run-parts can't just process
patches one at a time. Modified the build system to use the results of
run-parts --test and interate over the patches one by one, and take
action on a patch by patch basis.

Moved examples into /usr/share/doc/kernel-package/examples, as
suggested by policy. 


Handle the surces symlink just like we handle the build symlink,
complete with a new option.

Reverted the default for asking whether the failure of depmod was ok,
since it is indeed mostly fine.

At the request of Colin Watson, changed from find string iMac.  Any
powerpc system that would contain Mac in /proc/cpuinfo is a PowerMac
system, according to arch/ppc/platforms/* in the kernel source


Call the post-install hooks _after_ the image has been installed, not
before. Fixed a few typos in man pages.

git-archimport-id: srivasta@debian.org--2003-primary/kernel-package--devel--0.2--patch-48