Polishing the UML related code
More modifications to the kernel-uml-$version packages created. Now
the modules live in /usr/lib/modules/ instead of in
/usr/share/uml. The package now uses alternatives to allow people to
have multiple uml images on the machine at the same time, while still
retaining a /usr/bin/linux and a man page for that.  Rewrote the man
page from scratch.

The old postinst script for initrd kernels, in the case of an upgrade,
would overwrite the existing initrd in-place.  If there was a failure
for whatever reason (most notably, mkinitrd failing), the existing
initrd was destroyed and the system could be left in an unbootable
state in some cases where this could otherwise be prevented. Now, it
attempts to make the process somewhat more safe by writing to a
temporary file and atomically renaming it only if mkinitrd succeeds.


git-archimport-id: srivasta@debian.org--2003-primary/kernel-package--devel--0.2--patch-51