A major bug fix
make-kpkg: fails to unpatch the source, thanks to martin f
krafft. This was due to recent idiocy in upstream kernel sources,
where the kernel now assumes responsibility of ./debian directory
(despite Debian having used this directory for years). The people in
charge of this code in upstream kernel were not competent enough to
implement the deb target in a fashion that it does not break a
facility that Debian has used for years.  Well, all the kernel-patch
packages are affected; since nothing shall unpatch correctly with
2.6.8.1+ kerel versions. It also makes life a pain for official kernel
image developers, since their ./debian is a dorectory that is not
recreated by kernel-package (or the broken upstream code); and at
least one developer has lost an uncheked in custom ./debian dir

It is not as if they needed to claim ./debian, since they do not use
dpkg, and a fix was discussed in August, but no one has seen fit to
implement that fix and solve this breakage -- so now kerel-package
works around this idiocy. This makes this update very urgent.

The generated .debs also do not pay attention to kernel-img.conf (thus
creating a surprise for unsuspecting end users), they do not massage
the build and source symlinks, do not run boot loaders, do not handle
bootloaders that may use initrd in their config, and probably would
end up generating a slew of bugs that can't be fixed in Debian.
  

Bug fix: "make-kpkg doesn't handle CONFIG_LOCALVERSION", thanks to
Michal Čihař. The only implementation for LOCALVERSION I saw (in
linux-2.6.9-rc3/Makefile), is a broken Makefile
snippet. kernel-package now tries to implement what the documentation
and comment say should be done

Apart from hook variables that the local admin may set, there are a
set of directories where packages, or the local admin, may drop in
script files. The directories are

  - /etc/kernel/preinst.d/,
  - /etc/kernel/postinst.d/,
  - /etc/kernel/prerm.d/, and 
  - /etc/kernel/postrm.d/.

If they exists, the kernel-image package shall run a run-parts program
over the directory, giving the version being installed or removed as
an argument, in the corresponding phase of installation or
removal. Since these directories do not exist by default, this should
have no impact on current installations.

git-archimport-id: srivasta@debian.org--2003-primary/kernel-package--devel--0.2--patch-60