Updated mips rules, fixed problems with nuking scripts Makefiles
The was an unfortunate confluence of moving and touching the scripts
Makefile, and make clean: the dummy, empty Makefile was removed, and thus
never made it into the source package. The tarball seems to include
scripts/package/builddeb.dist and scripts/package/Makefile.dist instead
of scripts/package/builddeb and scripts/package/Makefile.  Fixed it to
echo a no-op comment into the dummy Makefile.

The current kernel-package has a rather limited understanding of
mips/mipsel subarchitectures, and fails to handle the 2.6/64bit
MIPS kernels which are currently in preparation. The appended
patch adds support for the currently projected subarchitectures.

For the kernels currently in the archive, nothing changes on the
mips side. On the mipsel side, the broken bootloader definitions
for everything which isn't a DECstation are fixed, and the xxs1500
image is changed from ELF to srec, which is what the firmware
understands.

git-archimport-id: srivasta@debian.org--2003-primary/kernel-package--devel--0.2--patch-74