A new bugfix release
We now mention that the generated silo.conf is a best effort kinda thing,
and a human needs to check it. Perhaps we should not even try to generate
boot loader files?

Simultaneously installing two kernel images, with initrds, makes lilo
fail for the first one", thanks to Adam Rice. Actually, this was an
instance of upgrading two images at the same time, which is significant,
since the fix really was in the last upload where we do not delete the
initrd symlink on upgrade.

git-archimport-id: srivasta@debian.org--2003-primary/kernel-package--devel--0.2--patch-78