Minor Fixes
Fixed the completions definition so it does not throw garbage on the
screen if ./config does not exist.

Fix the name of the file initrd.img is renamed to in case the old version
had an initrd, and the new version does not.

Fixed the tmpfile handling of kernel_grub_rm.sh.

git-archimport-id: srivasta@debian.org--2003-primary/kernel-package--devel--0.2--patch-83