New bug fixing version
Cleaned up the handling of the privilege acquisition program. We sued to
munge together the singing directives (-us and -uc) passed on to
dpkg-buildpackage, which led to some convoluted code, and did break some
dpkg-buildpackage invocation. Clarified the man page to elucidate the
separation of UNSIGN_SOURCE and UNSIGN_CHANGELOG from root_cmd. 

Added versioned sub directories to all the /etc/kernel/{pre,post}{inst,rm}.d/
directories, by popular demand. This allows the hook scripts to not have
to do version checks, and perhaps not run unrelated scripts on install,
at the cost of added directories and slight complexity increase in the
interface.  Documented the additional directories in the man page.

Upon maintainer request, do not run quik for official powerppc images.

Moved version determination code out to a makefile snippet, this allows
us to simplify the calculation of variables that are affected by the top
level Makefile (which includes things we put in conf.vars). 

Made the warning about initrd less intimidating, but have not yet removed
it entirely. This should be looked at again in Etch.

git-archimport-id: srivasta@debian.org--2003-primary/kernel-package--devel--0.2--patch-85