A workaround for header packages including build links
This is a work around for image packages being installed on a target
machine that has an official kernel headers package already installed,
and which thus provides build symbolic link. Now, official kernel-image
packages do not have a build symbolic link, which reverses what
kernel-package normally creates (image packages have build symbolic link,
header packages do not).

It would not be a good idea to just let official image packages ignore
the /lib/modules/$version existence check outright, since that means that
they may overwrite a kernel image that the user had with the same name.
A tight override is really needed -- only ignore the modules directory if
it contains a single build symbolic link, and nothing else.

This patch now includes a tight check in preinst, and if it is OK to
retain the build symbolic link, it is saved as build.save. In the
postinst, the build.save symbolic link is restored, iff the image package
did not itself install a build link (this allows a local image kernel to
be installed with the same name as a official kernel header package).


git-archimport-id: srivasta@debian.org--2003-primary/kernel-package--devel--0.2--patch-89