Fix dangling symlink checks in maintainer scripts
Several places in the maintainer scripts used the function readlink to
determine the target of a symbolic link without resolving it. However,
contrary to the expectation, this may result in the part removing the
dangling symlink never being triggered. There is now an extra check to
determine whether the target pointed to actually exists.

* kernel/image.postinst: Use Cwd from Perl-base for abs_path
  (fix_build_link):  Use abs_path to get to the real target, and test for
  existence. 
  (fix_source_link): Ditto
  (image_magic): Ditto (in a couple of separate places)

* kernel/image.postrm: Use Cwd from Perl-base for abs_path

* kernel/image.postrm (image_magic): Use abs_path to get to the real
  target, and test for existence.


git-archimport-id: srivasta@debian.org--etch/kernel-package--devel--9.0--patch-4