Circumvent version mismatch checks for official kernels
As all the debian kernel images are built from the same source, but with
different binary image versions (the binary image version is a
combination of the sub-arch and a variation of the kernel ABI), this check
is erroneously triggered for them, terminating the build. With the patch
builds finish fine.  This does so by circumventing a sanity check, which
has some value for the end user.


* kernel/rules (HAVE_VERSION_MISMATCH): Warn people about version
  mismatches, unless they are running an "official" version, in which
  case they can shoot themselves in the foot if they so desire.


git-archimport-id: srivasta@debian.org--etch/kernel-package--devel--9.0--patch-5