Undo the demangling of the Makefile in the scrpts dir
Since the scripts dir Makefile started nuking ./debian directories
willy-nilly, we have been moving it out of place and replacing it with an
innocuous replacement. Now that we are shipping the scripts dir in the
headers package, and since the headers package does not have a ./debian
directory we care about, we undo the move and ship the upstream
Makefile. 

* kernel/rules (real_stamp_headers): Undo the demangling of the Makefile
  in the scrpts dir, as

* make-kpkg (main): Since we modified the regular expression for the
  config target, we need to adjust the check as well.


git-archimport-id: srivasta@debian.org--etch/kernel-package--devel--9.0--patch-24