Try and make ./debian directory stand-alone
In the long run, make-kpkg should be a thin wrapper that merely populates
a few environment variables, and then calls make. One should be able to
just grab the ./debian directory, and be able to re-create the .deb
packages even on machines where kernel-package is not installed. Also
started on the first step towards making ./debian/rules more modular, by
eliminating many instances in which we were calling make recursively,
when a simple target dependency does the work.  This should allow us to
separate out the target dependencies required by the policy from the grun
work done to actually create the kernel-image package.

git-archimport-id: srivasta@debian.org--etch/kernel-package--devel--9.0--patch-44