Fix link manipulation when reinstalled after removal
The problem was that while the symbolic links were removed when the
package was removed, dpkg passed the last-version-configured to the
postinst, and we took that as evidence that the package had been
installed before -- which, while true, did not take into account that the
package was currently un-installed. The fix is to always see if a missing
symlink needs to be installed, and not touch existing symlinks.

* kernel/pkg/image/postinst (inird): call handle missing symlinks if needed.

* kernel/pkg/image/postinst (kimage): Ditto

* kernel/templates.in: Minor typo fix: $loader -> ${loader}

git-archimport-id: srivasta@debian.org--etch/kernel-package--devel--9.0--patch-57