Move build time hook invocations later into package creation
It was reported that the hooks called while various kernel packages were
being built were now called _before_ the maintainer scripts were put into
place, preventing any modifications of these scripts. In some cases, like
the image and headers packages, it was possible to move the hooks to a
point just before the package creation. In other cases, like for the
source package, this was not possible, since the source is tarred up
prior to packaging, and the hooks needs to be run before the tarring is
done. In the case of the manual package, there was no hook invoked.

* kernel/ruleset/targets/headers.mk (install/$(h_package)): Move the hook
  invocation away from the install target.

* kernel/ruleset/targets/headers.mk (debian/$(h_package)): And invoke it
  here instead.

* kernel/ruleset/targets/image.mk (install/$(i_package)): Move the clean
  hook out of here.

* kernel/ruleset/targets/image.mk (debian/$(i_package)): And into here.

git-archimport-id: srivasta@debian.org--etch/kernel-package--devel--9.0--patch-73