Allow building third party modules from header package directory
It was pointed out to me that the ability to build third party modules
using make-kpkg and linux kernel headers packages had broken with recent
kernel packages, since the check for whether we were in a top level
kernel source directory failed to take into account if we were in a
header package directory, which is sufficient.

Additionally, older kernel-package versions appear to want to run
oldconfig in the headers package tree, which failed, since there was
inadequate support for running oldconfig. This has also been fixed, since
all we really need to depend on is a .config file.

Finally, make-kpkg insists on there being a ./debian directory, and it
makes sense to ship the directory in the source and header packages,
since shipping it means that we also remove the need for kernel-package
to be installed on the target system

* kernel/ruleset/misc/checks.mk: The test for whether we are in a proper
  directory for running make-kpkg was incomplete, and failed to take into
  account if we were using a kernel headers package.

* kernel/ruleset/modules.mk (modules-image modules_image): Change the
  dependency from configure to .config (so as not to tun oldconfig)

* kernel/ruleset/modules.mk (modules): Ditto

* kernel/ruleset/modules.mk (modules-config modules_config): Ditto

* kernel/ruleset/modules.mk (modules-clean modules_clean): Ditto 

* kernel/ruleset/targets/source.mk (install/$(s_package)): Arrange for
  ./debian directory to be shipped in the package.

* kernel/ruleset/targets/headers.mk (install/$(h_package)): Arrange for
  ./debian directory to be shipped in the package.

git-archimport-id: srivasta@debian.org--etch/kernel-package--devel--9.0--patch-75