Add a new kernel image handling variable warn_boot.
Add a new kernel image handling variable called warn_reboot. This
variable can be used to turn off the warning given when installing a
kernel image which is the same version as the currently running
version. If the modules list is changed, the modules dependencies may
have been changed, and the modules for the new kernel may not run
correctly on the running kernel if the kernel ABI has changed in the
meanwhile. It is a good idea to reboot, and this is a note to remind
you. This variable is set by default.

* kernel-img.conf.5: Document the new warn_reboot variable, which reminds
  people to reboot.

* kernel/pkg/image/prerm: Add a new variable, warn_reboot

* kernel/pkg/image/postrm: Ditto

* kernel/pkg/image/preinst: Ditto

* kernel/pkg/image/postinst: Ditto

git-archimport-id: srivasta@debian.org--etch/kernel-package--devel--9.0--patch-184