Make all packages even for Xen and UML variants
Enable -source and -headers packages for Xen and UML. This is a simple
change, and should have minimal impact on any other package, or on
current user built packages. The upside is that the previously disabled
packages are enabled for Xen and UML variants, allowing people to build
third party modules for Xen adn UML variations -- and even if these were
to break (they do not, in my testing), there is no regression.  In the
future, the image and modules packages will be separated, to allow for
host/guest installs.


* kernel/ruleset/targets/doc.mk (install/$(d_package)): Make doc packages
  even for Xen and UML.
  (debian/$(d_package)): Ditto
  (binary/$(d_package)): Ditto

* kernel/ruleset/targets/headers.mk (install/$(h_package)): Make header
  packages even for Xen and UML
  (debian/$(h_package)): Ditto
  (binary/$(h_package)): Ditto

* kernel/ruleset/targets/manual.mk (install/$(m_package)): Make manual
  packages even for Xen and UML 
* kernel/ruleset/targets/manual.mk (debian/$(m_package)): Ditto.
* kernel/ruleset/targets/manual.mk (binary/$(m_package)): Ditto

* kernel/ruleset/targets/source.mk (install/$(s_package)): Make source
  packages even for Xen and UML.

* kernel/ruleset/targets/source.mk (debian/$(s_package)): Ditto

* kernel/ruleset/targets/source.mk (binary/$(s_package)): Ditto.



git-archimport-id: srivasta@debian.org--etch/kernel-package--devel--9.0--versionfix-1