Fix cut-and-paste error in maintainer scripts.
In kernel/pkg/image/p* the link_in_boot variable is set by incorrectly
checking for image_in_boot from the config file. This was an out-and-out
cut and paste error; and the fix is very simple (remove offending
line). The downside is minimal, and current implementation is just wrong.

* kernel/pkg/image/postinst: Fix cut-and-paste error

* kernel/pkg/image/postrm: Ditto.

* kernel/pkg/image/preinst: Ditto.

* kernel/pkg/image/prerm: Ditto.


git-archimport-id: srivasta@debian.org--etch/kernel-package--devel--9.0--versionfix-2