Update dependency on coreutils
In release version 10.063, I cleaned up the  postinst as while I was at
it, going with readlink -q -m instead of the built in version. Since -m
was a flag not present in Sarge's coreutils, this means the generated
image files should have a versioned dependency on coreutils. The
previous versioned dependency was not strict enough, updated to a
version that works.

* kernel/Control (Depends): We now depend on a coreutils that does
  contain a readlink -m flag.


git-archimport-id: srivasta@debian.org--etch/kernel-package--devel--9.0--versionfix-5