use DEB_HOST_ARCH_OS
kernel-package does not work on armel, as it uses DEB_HOST_GNU_SYSTEM and
compare it to either linux-gnu or kfreebsd-gnu. DEB_HOST_GNU_SYSTEM
equals to linux-gnueabi on armel.

The correct fix is to use DEB_HOST_ARCH_OS instead which returns linux on
armel as well as on other linux architectures. Closes: #425971

* kernel/ruleset/arches/amd64.mk (KERNEL_ARCH): use DEB_HOST_ARCH_OS

* kernel/ruleset/arches/i386.mk (DEBCONFIG): use DEB_HOST_ARCH_OS

* kernel/ruleset/minimal.mk (minimal_clean): use DEB_HOST_ARCH_OS

* kernel/ruleset/misc/defaults.mk: use DEB_HOST_ARCH_OS

* kernel/ruleset/misc/modules.mk: use DEB_HOST_ARCH_OS

* kernel/ruleset/misc/version_vars.mk: use DEB_HOST_ARCH_OS

* kernel/ruleset/targets/image.mk: use DEB_HOST_ARCH_OS 

* kernel/ruleset/targets/target.mk (conf.vars): use DEB_HOST_ARCH_OS 


git-archimport-id: srivasta@debian.org--lenny/kernel-package--devel--11.0--patch-14