add support for building lguest launcher when lguest host support is built-in or module
This commit uilds the lguest launcher when CONFIG_LGUEST is set
in the target kernel, and installs it to /lib/modules/{kernel-version}
in the linux-image package.

This is an appropriate way to do this, as lguest is tied quite strongly
to the version of the kernel it was built with (there's no defined ABI
or API, so it pretty much has to launch a kernel from the same source as
well, unless you're lucky), so there's no point in building a seperate
lguest package for the launcher.

However, this won't work when cross-building an i386 package (the only
one where vanilla kernels can have CONFIG_LGUEST defined right now)
because the lguest Makefile assumes that ld and gcc point to i386
versions. CONFIG_LGUEST is disabled in default configs.

Signed-off-by: Manoj Srivastava <srivasta@debian.org>