Move to a streamlined, make -j friendly build system
While I am not sure of this fixes some of the nagging problems we have
been facing in recent versions of kernel-package, where we used double
colon rules, which were convenient, sure, but played havoc with
ordering of the rules, and had to have various band-aids to help out
with the ordering. The system was rapidly growing complex, with clear
indication that it was actually faster.

The new target mechanism does away with doublecolon rules, and should
play better with parrallel compilation.

Signed-off-by: Manoj Srivastava <srivasta@debian.org>