Fix the build system to accommodate XEN changes, use kbuild as far as possible
These are fairly extensive changes. Firstly, we try to use upstream
kbuild as far as possible, to reduce churn as the files upstream
installs change. Some added checks of the Makefile are now in place so
we retain backwards compatibility. This should improve things lot wrt
header files.  We also now add dependencies to more packages actually
required to build kernel images.

We also try to look for the kbuild created KERNELRELEASE variable,
which is designed to be used by distros to figure out where modules
are to be loaded from, etc. This should help reduce version
mismatches. We also prepare the kernel.release file early, to help
that.

We also refitted to support the new XEN code in mainstream, in that
the same image can be booted normally or be used as a XEN image. This
support probably needs to be improved.

The make target dependencies have been extensively reworked, to
minimize surprises and wasted effort. We also strip modules, based on
DEB_BUILD_OPTIONS (nostrip).

Extra care is now taken so we do not accidentally remove ./debian
while cleaning, thanks to upstream helpfully removing ./debian when
cleaning.  This should prevent dpkg-buildpackage from accidentally
shoooting itself in the foot by removing ./debian as its first action.

Signed-off-by: Manoj Srivastava <srivasta@debian.org>