Modify docs now that make-kpkg clean is not a requirement
The new build system is far more robust about keeping up with version
numbers, and so all that is needed to change a .config or a version
number is to just rm -rf ./debian and start building. This cuts down a
lot on time wasted building kernels and deleting the .o files before
the next build.

Signed-off-by: Manoj Srivastava <srivasta@debian.org>